requirements.txt
Django==2.0.5
matplotlib==2.2.2
numpy==1.14.3
pandas==0.23.0
Pillow==5.1.0
teamcity-messages==1.21
